% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectioncommentAddin.R
\name{comment_styler}
\alias{comment_styler}
\title{Function to create the padded comment}
\usage{
comment_styler(x, double_mode=FALSE, l = 80)
}
\arguments{
\item{x}{a string to format}

\item{double_mode}{a logical indicating only to convert 
lines starting with at least a double hash sign}

\item{l}{an integer indicating length to pad}
}
\value{
a string that is formatted
}
\description{
This function is the logic that runs to reformat
the comment as a section break
}

