% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectioncomment.R, R/sectioncommentAddin.R
\docType{package}
\name{sectioncomment}
\alias{sectioncomment}
\alias{sectioncomment-package}
\title{sectioncomment}
\usage{
sectioncomment
}
\description{
See the full \href{https://github.com/reportmort/sectioncomment}{README} on GitHub.

Creates an 80 character width comment padded with hyphens, similar to how 
comments are shown in Hadley's style guide (http://adv-r.had.co.nz/Style.html)
}

